+++ esphome/components/emc2101/emc2101.cpp
@@ -161,39 +161,6 @@ float Emc2101Component::get_speed_() {
   return tach == 0xFFFF ? 0.0f : 5400000.0f / tach;
 }
 
-void EMC2101Output::write_state(float state) { this->parent_->set_duty_cycle_(state); }
-
-float EMC2101Sensor::get_setup_priority() const { return setup_priority::DATA; }
-
-void EMC2101Sensor::dump_config() {
-  ESP_LOGCONFIG(TAG, "Emc2101 sensor:");
-  LOG_SENSOR("  ", "Internal temperature", this->internal_temperature_sensor_);
-  LOG_SENSOR("  ", "External temperature", this->external_temperature_sensor_);
-  LOG_SENSOR("  ", "Speed", this->speed_sensor_);
-  LOG_SENSOR("  ", "Duty cycle", this->duty_cycle_sensor_);
-}
-
-void EMC2101Sensor::update() {
-  if (this->internal_temperature_sensor_ != nullptr) {
-    float internal_temperature = this->parent_->get_internal_temperature_();
-    this->internal_temperature_sensor_->publish_state(internal_temperature);
-  }
-
-  if (this->external_temperature_sensor_ != nullptr) {
-    float external_temperature = this->parent_->get_external_temperature_();
-    this->external_temperature_sensor_->publish_state(external_temperature);
-  }
-
-  if (this->speed_sensor_ != nullptr) {
-    float speed = this->parent_->get_speed_();
-    this->speed_sensor_->publish_state(speed);
-  }
-
-  if (this->duty_cycle_sensor_ != nullptr) {
-    float duty_cycle = this->parent_->get_duty_cycle_percent_();
-    this->duty_cycle_sensor_->publish_state(duty_cycle);
-  }
-}
 
 }  // namespace emc2101
 }  // namespace esphome
