+++ esphome/components/emc2101/emc2101.h
@@ -1,10 +1,7 @@
 #pragma once
 
-#include "esphome/core/component.h"
-#include "esphome/core/helpers.h"
-#include "esphome/components/output/float_output.h"
-#include "esphome/components/sensor/sensor.h"
 #include "esphome/components/i2c/i2c.h"
+#include "esphome/core/component.h"
 
 namespace esphome {
 namespace emc2101 {
@@ -118,45 +115,6 @@ class Emc2101Component : public Component, public i2c::I2CDevice {
   Emc2101DACConversionRate dac_conversion_rate_;
 };
 
-/// This class allows to control the EMC2101 output.
-class EMC2101Output : public output::FloatOutput {
- public:
-  EMC2101Output(Emc2101Component *parent) : parent_(parent) {}
-
- protected:
-  /** Used by ESPHome framework. */
-  void write_state(float state) override;
-
-  Emc2101Component *parent_;
-};
-
-/// This class exposes the EMC2101 sensors.
-class EMC2101Sensor : public PollingComponent {
- public:
-  EMC2101Sensor(Emc2101Component *parent) : parent_(parent) {}
-  /** Used by ESPHome framework. */
-  void dump_config() override;
-  /** Used by ESPHome framework. */
-  void update() override;
-  /** Used by ESPHome framework. */
-  float get_setup_priority() const override;
-
-  /** Used by ESPHome framework. */
-  void set_internal_temperature_sensor(sensor::Sensor *sensor) { internal_temperature_sensor_ = sensor; }
-  /** Used by ESPHome framework. */
-  void set_external_temperature_sensor(sensor::Sensor *sensor) { external_temperature_sensor_ = sensor; }
-  /** Used by ESPHome framework. */
-  void set_speed_sensor(sensor::Sensor *sensor) { speed_sensor_ = sensor; }
-  /** Used by ESPHome framework. */
-  void set_duty_cycle_sensor(sensor::Sensor *sensor) { duty_cycle_sensor_ = sensor; }
-
- protected:
-  Emc2101Component *parent_;
-  sensor::Sensor *internal_temperature_sensor_{nullptr};
-  sensor::Sensor *external_temperature_sensor_{nullptr};
-  sensor::Sensor *speed_sensor_{nullptr};
-  sensor::Sensor *duty_cycle_sensor_{nullptr};
-};
 
 }  // namespace emc2101
 }  // namespace esphome
