+++ esphome/components/emc2101/output/__init__.py
@@ -2,17 +2,15 @@ import esphome.codegen as cg
 import esphome.config_validation as cv
 from esphome.components import output
 from esphome.const import CONF_ID
-from . import Emc2101Component, emc2101_ns
+from .. import EMC2101_COMPONENT_SCHEMA, CONF_EMC2101_ID, emc2101_ns
 
 DEPENDENCIES = ["emc2101"]
 
-EMC2101Output = emc2101_ns.class_("EMC2101Output", output.FloatOutput, cg.Component)
-CONF_EMC2101_ID = "emc2101_id"
+EMC2101Output = emc2101_ns.class_("EMC2101Output", output.FloatOutput)
 
-CONFIG_SCHEMA = output.FLOAT_OUTPUT_SCHEMA.extend(
+CONFIG_SCHEMA = EMC2101_COMPONENT_SCHEMA.extend(
     {
         cv.Required(CONF_ID): cv.declare_id(EMC2101Output),
-        cv.GenerateID(CONF_EMC2101_ID): cv.use_id(Emc2101Component),
     }
 )
